requirements.txt
streamlit==1.28.0
openai==1.2.4
pandas==2.1.1
numpy==1.26.1
pymongo==4.5.0
python-dotenv==1.0.0
scikit-learn==1.3.2
nltk==3.8.1
plotly==5.17.0
pydantic==2.4.2
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
tqdm==4.66.1
pytest==7.4.3
faiss-cpu==1.7.4
networkx==3.2.1